Rename files and add clang_ast_to_asr.cpp
Which contains all the clang includes
